requirements.txt
# Core dependencies
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
tqdm>=4.64.0
openai-agents>=0.2.7

# Data processing
typing-extensions>=4.0.0

# Development dependencies (optional)
jupyter>=1.0.0

